wordriver plugin: add ID property
Previous iteration had an odd level of specificity, and created a
single div in any given target for the wordriver.  By adding an ID, we
can create one or multiple wordriver instantiations in the target.

I *hope* this works OK. I'm a little worried that this new parent will take up
space even when the contents are hidden.